refactor(frontend): split Navbar into navbar/UserDropdown + NavbarBrand
Phase 5.5 — no-behavior/no-visual split (1731 -> 1294):
- navbar/UserDropdown.svelte (487) — user avatar/button + dropdown menu. Preserves the
  EXACT .user-button/.dropdown-menu/.dropdown-item DOM + classes the settings E2E drives;
  kept the original bubble-phase click-outside (not switched to the capture-phase
  clickOutside action, to avoid timing differences); exposes close() for the parent's
  notifications-panel coupling; events open/openSettings/itemSelected/logout.
- navbar/NavbarBrand.svelte (74) — logo/branding; event 'about'.
Navbar keeps top-level layout, routing/active-link, store subscriptions, and the
recording-indicator (tightly coupled to ~7 handlers + its local always-padded formatDuration
— intentionally not extracted/migrated). Cross-boundary mobile rules converted to
.nav-links.open :global(...) (the existing pattern). Avatar markup unchanged (single-letter,
not ui/Avatar which renders 2-char initials).

Verified: svelte-check 0/0 (861 files), build, 52 unit tests, settings E2E 3/3 (navbar user
menu -> Settings works).
